source/requirements.txt
Pillow
numpy

matplotlib
pandas
